% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{cindex_times}
\alias{cindex_times}
\alias{cindex_wrapper}
\title{Returns (apparent) C-Index for each time point specified in \code{times}.}
\usage{
cindex_times(object, newdata, times, form.ipcw = Hist(survtime, event) ~ 1,
  cens.model = "none", ...)

cindex_wrapper(model, data, times = NULL, form.ipcw = Hist(survtime, event)
  ~ 1, cens.model = "none", tvar = "tend", ...)
}
\arguments{
\item{object}{A matrix with survival probabilities (rows = patients, cols =
time).}

\item{newdata}{Data set on which C-Index will be evaluated.}

\item{times}{Times at which C-Index should be evaluated.}

\item{form.ipcw}{Formula for the model which will be used to calculate inverse
probabilities of censoring weights.}

\item{cens}{model If "none", no IPCW correction is performed.}
}
\description{
Returns (apparent) C-Index for each time point specified in \code{times}.
}
