% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{cindex_t}
\alias{cindex_t}
\title{Calculate C-Index at time t}
\usage{
cindex_t(pmat, newdata, t, form.ipcw = Hist(survtime, event) ~ 1,
  cens.model = "none", ...)
}
\arguments{
\item{pmat}{Matrix of surival probabilities (rows = patients, cols = time}

\item{newdata}{For the apparent C-Index same data that was used to calculate
\code{pmat}.}

\item{t}{Time point at which C-Index should be evaluated.}

\item{form.ipcw}{Formula for the model which will be used to calculate inverse
probabilities of censoring weights.}

\item{cens}{model If "none", no IPCW correction is performed.}
}
\description{
For specified time t and given survival probabilities for this time point,
retuns (apparent) C-index evaluated at the respective time t.
}
