% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotComparisons.R
\name{ggplot_comparison}
\alias{ggplot_comparison}
\alias{ggbox_comparisons}
\alias{gglines_comparisons}
\alias{ggbox_comparison_diffs}
\alias{gglines_comparison_diffs}
\alias{ggtext_simstats}
\alias{ggcomparison_lines2}
\title{Plot functions for hazard ratio comparisons given different nutrition profiles}
\usage{
ggplot_comparison(comparisons, true.comparisons = NULL, ...)

ggbox_comparisons(gg.base.comparison, ...)

gglines_comparisons(gg.base.comparison, alpha = 0.1, ...)

ggbox_comparison_diffs(gg.base.comparison, ...)

gglines_comparison_diffs(gg.base.comparison, alpha = 0.1, ...)

ggtext_simstats(simstats, id.vars = c("label", "comparison"), ...)

ggcomparison_lines2(setting, alpha = 0.1)
}
\arguments{
\item{comparisons}{A data frame defining the comparisons of interest.}

\item{true.comparisons}{If provided, these true comparisons will be added to
the plot of simulated comparisons}
}
\description{
\code{ggplot_comparison} is the base function. Other functions mostly extensions.
}
\keyword{internal}
