% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLmat.R
\name{create_Lmat}
\alias{create_Lmat}
\title{creates one instance of Lag/Lead mat}
\usage{
create_Lmat(te = 1:12, te.add.lag = 0, t.lag = 4, te.add.lead = 0,
  lead.const = 4, lead.factor = 2, interval.seq = c(0:12, seq(15, 55, by =
  5), 61.1), t.min = 0)
}
\arguments{
\item{te}{Numeric/Integer vector specifying the times at which exposure occurred.}

\item{te.add.lag}{A numeric constant added to te before application of lag time}

\item{t.lag}{A numeric constant, specifying the time (from \code{te}) before
\code{te} can affect hazard.}

\item{te.add.lead}{A numeric constant, added to te before application of lead time.}

\item{lead.const}{A numeric constant, specifying the constant amount of time
after \code{te + t.lag}, in which \code{te} can still affect hazard.}

\item{lead.factor}{If the lead time is dynamic, this factor can be set different
to zero, such that \code{t.lead=lead.const + lead.factor*te}.}

\item{interval.seq}{The break points dividing the follow up into intervals.}

\item{t.min}{If some intervals are not of interest only intervals for t > t.min are
returned.}
}
\value{
A data frame with intervals as first column and \code{length(te)}
columns specifying the lag/lead for each \code{te}.
}
\description{
creates one instance of Lag/Lead mat
}
\keyword{internal}
