% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNutritionData.R
\name{createNutritionVars}
\alias{createNutritionVars}
\title{Creates nutrition adequacy variables and Lag/Lead matrices}
\usage{
createNutritionVars(valid.cases, brks = c(0:12, seq(15, 55, by = 5), 61.1),
  ignore.first = TRUE, vars.ignore = NULL)
}
\arguments{
\item{valid.cases}{Data set in piece-wise exponential data format.}

\item{brks}{Breaks used to cut the time line for piece-wise exponential format.}

\item{ignore.first}{Indicates whether first observation of the TDC}

\item{vars.ignore}{List of variable names specifying TDC for which first
column should be removed.}
}
\description{
Creates nutrition adequacy variables and Lag/Lead matrices
}
\keyword{internal}
