% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-model-evaluation.R
\name{eval_model}
\alias{eval_model}
\alias{get_comparison_diffs}
\alias{get_df}
\alias{get_custom_df}
\alias{get_rdf}
\alias{test_effect_diffs}
\alias{get_comparison_pvals}
\alias{get_effect_diffs}
\alias{proto_lp}
\alias{get_true_diffs}
\alias{get_X}
\alias{melt_comparisons}
\alias{add_truth}
\title{Functions for evaluation of Simulation Study Part A.}
\usage{
eval_model(pam, X.list, protocols.to.compare, protocols.df, maxdays.tdc, intmid,
  debug = TRUE, ...)

get_comparison_diffs(coef.vec, Vp, X.list, protocols.to.compare, protocols.df,
  maxdays.tdc, intmid)

get_df(model, useColumns)

get_custom_df(edf, X1, X2)

get_rdf(model)

test_effect_diffs(model, X1, X2, effectname = "AdequacyCals", debug = FALSE,
  debugX = FALSE)

get_comparison_pvals(model, X.list, protocols.df, protocols.to.compare,
  maxdays.tdc, debug = FALSE)

get_effect_diffs(coef.vec, Vp, X1, X2, intmid, effectname = "AdequacyCals",
  alpha = 0.05, debug = FALSE)

proto_lp(pattern, model, ped, median.x, effectname = "AdequacyCals",
  time.var = "intmid", id.var = "CombinedID", type = "lpmatrix",
  se.fit = FALSE)

get_true_diffs(static, truth.pattern = "Expertt1", rename.cols = c("fit",
  "se", "hi", "lo"))

get_X(Xcompare.list, truth.pattern = "Expertt1")

melt_comparisons(comparisons.list, id.vars = c("fit", "intmid", "se", "hi",
  "lo", "label"))

add_truth(comparisons.df, true.comparisons)
}
\arguments{
\item{pam}{A piece-wise exponential model (gam).}

\item{X.list}{List of model matirces representing protocol Comparisons.}

\item{protocols.to.compare}{Data frame of protocols that should be compared}

\item{protocols.df}{Dayly protocol regimen for protocols defined in \code{protocols.to.compare}.}

\item{maxdays.tdc}{Maximal number of days nutrition protocol was recorded.}

\item{intmid}{Interval midpoints of intervals in which follow-up is devided in.}

\item{...}{further arguments, currently not used.}

\item{X1}{Design matrix.}

\item{X2}{Design matrix.}

\item{effectname}{Name of variable(s) for which the effect should be extracted.
\code{"AdequacyCals"} or \code{"ProtCat"}.}
}
\value{
List containing mdoel coefficients and their covariance, summary of
the model object, Estimated ln HR of the comparisons and p-values for each
comparison.
}
\description{
The main function \code{eval_model} is called in the wrapper algorithm after
model is fit. Other functions basically helper functions.
}
\keyword{internal}
