% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_mgcv_teststat.R
\name{get_comp_diffs}
\alias{get_comp_diffs}
\title{get differences of effects on log hazard}
\usage{
get_comp_diffs(patients.list, protocols.df, protocols.to.compare, model,
  effectname = "AdequacyCals")
}
\description{
get differences of effects on log hazard
}
\keyword{internal}
