% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchJobHelpers.R
\name{get_pattern}
\alias{get_pattern}
\alias{get_maxIterations_perNode}
\alias{get_runtime_perNode}
\alias{get_convergence_info}
\title{Extracts string from simulation log files, following a specified pattern.}
\usage{
get_pattern(reg, ids, pattern = ".*Algorithm parameters.*form=*")

get_maxIterations_perNode(reg, ids)

get_runtime_perNode(reg, ids)

get_convergence_info(reg, ids)
}
\arguments{
\item{reg}{A registry object (see BatchJobs documentation)}

\item{ids}{A subset of job ids}

\item{pattern}{Greps Log files for Lines with this pattern.}
}
\description{
Extracts string from simulation log files, following a specified pattern.

Extract number of iterations from Log files

Extract runtime per job

Wrapper function that obtains iterations, convergence, runtime
and node name for specific Jobs (\code{ids}).
}
\keyword{internal}
