% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{get_hazard}
\alias{get_hazard}
\title{Returns data frame containing predicted hazards and selected variables}
\usage{
get_hazard(object, newdata, t, start.var = "tstart",
  vars.keep = c("CombinedID", "tstart", "tend", "survtime"))
}
\description{
Returns data frame containing predicted hazards and selected variables
}
