% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{get_iddf}
\alias{get_iddf}
\title{Given data set in long format, returns data set of last observations for an}
\usage{
get_iddf(df, t, id.var = "CombinedID")
}
\description{
Given data set in long format, returns data set of last observations for an
}
