% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{get_survprob}
\alias{get_survprob}
\title{Transforms a data frame containing Survival probabilities in long format to}
\usage{
get_survprob(St.df)
}
\description{
Transforms a data frame containing Survival probabilities in long format to
}
