% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{get_tid}
\alias{get_tid}
\title{Select the interval in which given t falls and return according data set, id
or index}
\usage{
get_tid(df, t, return.val = "df", id.var = "CombinedID",
  start.var = "tstart", end.var = "tend", select = NULL)
}
\description{
Select the interval in which given t falls and return according data set, id
or index
}
