% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNutritionEffect.R
\name{getEffectDiffs}
\alias{getEffectDiffs}
\title{Extract effect differences for 2 hypothetical patients}
\usage{
getEffectDiffs(m, newdata1, newdata2, effectname = "AdequacyCals")
}
\arguments{
\item{effectname}{Name of variable(s) for which the effect should be extracted.}
}
\description{
Extract effect differences for 2 hypothetical patients
}
\keyword{internal}
