% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPED.R
\name{getIntmid}
\alias{getIntmid}
\title{Simulate Data for simulation Part A}
\usage{
getIntmid(brks, seq.obs)
}
\description{
Draws survival times from Piece-wise Exponential distribution and
return new data set in Piece-wise Exponential data (PED) format.
}
