% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLmat.R, R/plot-lag-lead.R
\name{heatAdequacy}
\alias{heatAdequacy}
\alias{heatAdequacy}
\title{heatmap of the effect terms for relevant L-Areas}
\usage{
heatAdequacy(hm = NULL, day.names = 1:11, int.names = NULL,
  high.col = "steelblue", low.col = "white", grid.col = "lightgrey",
  title.char = "")

heatAdequacy(hm = NULL, day.names = 1:11, int.names = NULL,
  high.col = "steelblue", low.col = "white", grid.col = "lightgrey",
  title.char = "")
}
\arguments{
\item{hm}{A lag-lead matrix}

\item{day.names}{Labels applied to exposure times}

\item{int.names}{Labels applied to intervals in which follow-up was partitioned}

\item{high.col}{Color used to highlight exposure times within the lag-lead window}

\item{low.col}{Color of exposure times outside the lag-lead window.}

\item{grid.col}{Color of grid lines.}

\item{title.char}{Will be added at plot title}
}
\description{
heatmap of the effect terms for relevant L-Areas

Given a matrix defining a lag lead window, returns respective plot as a
\code{ggplot2} object.
}
\keyword{internal}
