% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{int_info}
\alias{int_info}
\title{create start/end times, given breaks}
\usage{
int_info(brks = c(0:12, seq(15, 55, by = 5), 61), min.int = 4L)
}
\description{
create start/end times, given breaks
}
\keyword{internal}
