% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLmat.R
\name{createLmatDyn}
\alias{createLmatDyn}
\alias{t_lead}
\alias{lag_lead_df}
\alias{get_end}
\alias{get_interval}
\alias{int_info2}
\title{Helper functions for creation of Lag-Lead matrices}
\usage{
createLmatDyn(lead = 4, lag.c = 4, lag.f = 2, n.days = 11,
  brks = c(0:11, seq(15, 55, by = 5), 61.1), time.shift = 4)

t_lead(te, te.add.lead = 0, lead.const = 4, lead.factor = 2)

lag_lead_df(te = 1:12, te.add.lag = 0, t.lag = 4, te.add.lead = 0,
  lead.const = 4, lead.factor = 2, interval.seq = c(0L:12L, seq(15, 55, by
  = 5), 61), labels = TRUE)

get_end(start, lead, max.end)

get_interval(x, interval.seq = c(0L:12L, seq(15, 55, by = 5), 61),
  labels = FALSE)

int_info2(brks = c(0:12, seq(15, 55, by = 5), 61), min.int = 4)
}
\description{
Helper functions for creation of Lag-Lead matrices
}
\keyword{internal}
