% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{last_ind}
\alias{last_ind}
\title{Given data set in long format and ordered by an id, returns last}
\usage{
last_ind(df, id.var = "CombinedID")
}
\description{
Given data set in long format and ordered by an id, returns last
}
