% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLmat.R
\name{Lmat_data}
\alias{Lmat_data}
\title{creates Lag/Lead matrix for all observations in a data set}
\usage{
Lmat_data(data, Lmat, merge.col.data = "int", merge.col.Lmat = "intsL")
}
\arguments{
\item{data}{The complete data set.}

\item{Lmat}{A data frame where first column specifies intervals and the other}

\item{merge.col.data}{The name of the column (in data) on which data and Lmat}

\item{merge.col.Lmat}{The name of the column (in Lmat) on which data and Lmat
should be merged. Defaults to \code{"intsL"}}
}
\description{
creates Lag/Lead matrix for all observations in a data set
}
\details{
:
Lmat, that only contains information for all unique intervals and all days of
exposure, is merged with data, such that correct rows of Lmat are added to the
data, but only the Lmat matrix (now with as many rows as data) is returned.
}
\keyword{internal}
