% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_mgcv_teststat.R
\name{median_patient}
\alias{median_patient}
\title{Extract median covariate information from columns}
\usage{
median_patient(data, patient.vars = c("ApacheIIScore", "Age", "BMI", "Year",
  "DiagID2", "AdmCatID", "Gender", "freqInMVd2to4", "freqPFd2to4",
  "freqOId2to4", "freqPNd2to4"), random.effect = "CombinedicuID",
  re.byvar = "icuByDummy")
}
\description{
Extract median covariate information from columns
}
\keyword{internal}
