% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformToPEM.R
\name{preprocess_for_ped}
\alias{preprocess_for_ped}
\title{Transform data in long format to wide format}
\usage{
preprocess_for_ped(data.long, patient, ICU, timevarying,
  timevar = "Study_Day", idvar = "CombinedID",
  calSurvVar = "calendarSurvdays")
}
\description{
Transform data in long format to wide format
}
\keyword{internal}
