% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPED.R
\name{sim_ped}
\alias{sim_ped}
\alias{getAdequacyMat}
\alias{getLmat}
\alias{getOffset}
\alias{complete_data}
\title{Simulate Data for simulation Part A}
\usage{
sim_ped(hazards.ij, complete.data, time.shift = 4, brks = c(0:12, seq(15,
  55, by = 5), 61.1), ints = c(4:12, seq(15, 55, by = 5), 61),
  vars.to.update = c("CombinedID", "intmid", "offset", "PatientDied", "event",
  "Survdays", "intMat"), vars.to.delete = c("inICU", "inMechVent"),
  re = TRUE, survvar = "Survdays")

getAdequacyMat(mat, ind.first, n.obs.old, n.obs.new)

getLmat(mat, ind.first, n.obs.old, n.obs.new)

getOffset(brks, n.obs, seq.obs, surv)

complete_data(m, train, time.shift = 4, brks = c(0:12, seq(15, 55, by = 5),
  61.1), ints = c(4:12, seq(15, 55, by = 5), 61),
  vars.to.update = c("CombinedID", "intmid", "offset", "PatientDied", "event",
  "Survdays", "intMat"), vars.to.delete = c("inICU", "inMechVent"),
  hazards = NULL)
}
\arguments{
\item{hazards.ij}{Hazard for each subject and interval combination.}

\item{complete.data}{Full data set (i.e. with one row per interval)
for each subject and complete exposure history in PED format.}
}
\description{
Draws survival times from Piece-wise Exponential distribution and
return new data set in Piece-wise Exponential data (PED) format.
}
\keyword{internal}
