% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureFunctions.R
\name{mse}
\alias{mse}
\alias{mae}
\alias{getFreqOrigInSim}
\alias{getFreqSimInOrigKI}
\alias{get_stats}
\alias{get_stats2}
\alias{get_valid_ids}
\alias{get_experiments}
\alias{get_all_comparisons}
\alias{get_all_pvals}
\title{Functions for simulation evaluation measures calculation}
\usage{
mse(dif)

mae(dif)

getFreqOrigInSim(df.fit)

getFreqSimInOrigKI(df.fit)

get_stats(comparisons.diff.df, ci.level = 0.05)

get_stats2(comparisons.diff.df, ci.level = 0.05)

get_valid_ids(reg, prob.pattern, algo.pattern, algo.form,
  exclude.error = TRUE, exclude.warning = FALSE)

get_experiments(reg, probs, algos, forms, min.n)

get_all_comparisons(reg, mc.cores = 1, probs = NULL, algos = NULL,
  forms = NULL, min.n = 1, settings = NULL, exclude.error = TRUE,
  exclude.warning = TRUE)

get_all_pvals(reg, mc.cores = 1, probs = NULL, algos = NULL,
  forms = NULL, min.n = 1, settings = NULL, exclude.error = TRUE,
  exclude.warning = TRUE)
}
\description{
Functions for simulation evaluation measures calculation

extract comparison measures from data frame with comparisons differences from
simulations and true effects/true differences

Wrapper functions that obtains evaluated comparisons for jobs meeting
some criteria
}
\keyword{internal}
