% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{test_monotony}
\alias{test_monotony}
\title{Tests if provided survival probabilities are monotnously decreasing.}
\usage{
test_monotony(survprobs)
}
\description{
Tests if provided survival probabilities are monotnously decreasing.
}
