% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{test_survprob}
\alias{test_survprob}
\title{Given a matrix of survival probabilities (rows = ids, cols = times), tests}
\usage{
test_survprob(pmat, t)
}
\description{
Given a matrix of survival probabilities (rows = ids, cols = times), tests
}
