% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformToPEM.R
\name{transform_to_ped}
\alias{transform_to_ped}
\title{transforms data in long format into data in ped format (Piece-wise
exponential data)}
\usage{
transform_to_ped(data.long, timevarying, patient, ICU, brks,
  idvar = "CombinedID", timevar = "Study_Day", min.day = 0,
  survvar = "Survdays", calSurvVar = "calendarSurvdays", max.follow = 30)
}
\description{
transforms data in long format into data in ped format (Piece-wise
exponential data)
}
\keyword{internal}
