% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformToPEM.R
\name{wide_to_ped}
\alias{wide_to_ped}
\title{transform data in wide format back to long/ped format}
\usage{
wide_to_ped(data.wide, timevarying, brks = c(0:12, seq(15, 55, by = 5), 61.1),
  idvar = "CombinedID", timevar = "Study_Day", survvar = "Survdays",
  max.follow = 30)
}
\description{
transform data in wide format back to long/ped format
}
\keyword{internal}
