% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLmat.R
\name{heatAdequacy}
\alias{heatAdequacy}
\title{heatmap of the effect terms for relevant L-Areas}
\usage{
heatAdequacy(hm = NULL, day.names = 1:11, int.names = NULL,
  high.col = "steelblue", low.col = "white", grid.col = "lightgrey",
  title.char = "")
}
\description{
heatmap of the effect terms for relevant L-Areas
}
\keyword{internal}
